+++ Configuration/Icons.php
@@ -1,8 +1,0 @@
-<?php
-
-return [
-    'changelog_info' => [
-        'provider' => \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
-        'source' => 'EXT:changelog_info/Resources/Public/Icons/book-solid.svg',
-    ],
-];