+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Improve README
+- Make module available via "Web > Info > Changelog information" instead
+  of own main backend module.
 
 ## [1.1.1] - 2024-12-16